backend_fast/app/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
